file-parser-api/requirements.txt.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0
pandas==2.1.3
openpyxl==3.1.2
pdfplumber==0.10.3